+++ curvine-cli/src/cmds/fs/mod.rs
@@ -1,4 +1,16 @@
+mod cat;
 mod commands;
+mod common;
 mod content_summary;
+mod count;
+mod df;
+mod du;
+mod get;
+mod ls;
+mod mkdir;
+mod put;
+mod rm;
+mod stat;
+mod touch;
 
 pub use commands::FsCommand;
